reduce CompTimeScrubber deps and move renderCompTimeScrubber into own fn
