+++ src/composition/timeline/scrubber/CompTimeScrubber.styles.ts
